Improve handling of auto-delete with multiple instances
This commit improves the handling of Thumbnailer.lua's auto-delete
thumbnail cache feature when there are multiple instances of the same
video playing. Basically, a lock will be created in the thumbnail folder
when a video is playing. When the video finishes playing, the lock will
be released and the thumbnails folder is deleted as usual.

Assuming both videos are the same file and auto-delete is enabled, this
fix prevents the 2nd video from losing its thumbnails when the 1st video
is closed.

Relevant issues: deus0ww/mpv-conf#15 hooke007/MPV_lazy#48
Original commit: hooke007/MPV_lazy@273969f
Credit goes to the original author, of course.

+++ scripts/Thumbnailer.lua
@@ -180,6 +180,31 @@ local function delete_dir(path)
     return run_subprocess( OPERATING_SYSTEM == OS_WIN and {'cmd', '/e:on', '/c', 'rd', '/s', '/q', path} or {'rm', '-r', path} )
 end
 
+local function delete_file(path)
+    if not file_exists(path) then return end
+    msg.warn('Deleting File:', path)
+    return os.remove(path)
+end
+
+local function add_lock(path)
+    msg.debug('Add file lock to:', path)
+    local file = io.open(join_paths(path, tostring(utils.getpid())), 'w')
+    if file then
+        file:close()
+        return true
+    end
+    return false
+end
+
+local function remove_lock(path)
+    msg.debug('Remove file lock from:', path)
+    return delete_file(join_paths(path, tostring(utils.getpid())))
+end
+
+local function is_locked(path)
+    return #utils.readdir(path,'files') ~= 0
+end
+
 
 --------------------
 -- Data Structure --
@@ -411,6 +436,7 @@ end
 local stop_conditions
 
 local worker_script_path
+local auto_delete = nil
 
 local function create_workers()
     local workers_requested = (state and state.max_workers) and state.max_workers or user_opts.max_workers
@@ -460,6 +486,12 @@ local function create_ouput_dir(filepath, filename, dimension, rotate)
         return {basepath = nil, fullpath = nil}
     end
     msg.debug('Creating Output Dir: Using ', name)
+
+    if auto_delete == nil then auto_delete = user_opts.auto_delete end
+    if auto_delete > 0 then
+        add_lock(user_opts.cache_dir)
+        add_lock(basepath)
+    end
 
     local fullpath = join_paths(basepath, dimension, rotate)
     if not create_dir(fullpath) then return { basepath = nil, fullpath = nil } end
@@ -620,26 +652,36 @@ local function is_thumbnailable()
     return true
 end
 
-local auto_delete = nil
-
 local function delete_cache_dir()
+    local path = user_opts.cache_dir
+    remove_lock(path)
+
     if auto_delete == nil then auto_delete = user_opts.auto_delete end
     if auto_delete > 0 then
-        local path = user_opts.cache_dir
-        msg.debug('Clearing Cache on Shutdown:', path)
-        if path:len() < 16 then return end
-        delete_dir(path)
+        if not is_locked(path) then
+            msg.debug('Clearing Cache on Shutdown:', path)
+            if path:len() < 16 then return end
+            delete_dir(path)
+        else
+            msg.debug('Clearing Cache on Shutdown:ignore ', path, '- Locked')
+        end
     end
 end
 
 local function delete_cache_subdir()
     if not state then return end
+    local path = state.cache_dir_base
+    remove_lock(path)
+
     if auto_delete == nil then auto_delete = user_opts.auto_delete end
     if auto_delete == 1 then
-        local path = state.cache_dir_base
-        msg.debug('Clearing Cache for File:', path)
-        if path:len() < 16 then return end
-        delete_dir(path)
+        if not is_locked(path) then
+            msg.debug('Clearing Cache for File:', path)
+            if path:len() < 16 then return end
+            delete_dir(path)
+        else
+            msg.debug('Clearing Cache for File:ignore ', path, '- Locked')
+        end
     end
 end
